throwing and falling fixes for items and units
- when unit falls on another unit it now goes unconscious as per vanilla
behavior
- added stand up animations (from downed state)
- fixed multiple bugs with throwing in the air and flying
- proper fall speed for units and items
- proper delay after throw (as per vanilla behavior)

+++ game/state/battle/battledoor.cpp
@@ -78,9 +78,9 @@ void BattleDoor::update(GameState &, unsigned int ticks)
 	// Update animation
 	if (animationTicksRemaining > 0)
 	{
-		if (animationTicksRemaining > ticks)
+		if (animationTicksRemaining > (int)ticks)
 		{
-			animationTicksRemaining -= ticks;
+			animationTicksRemaining -= (int)ticks;
 		}
 		else
 		{
